Moved ChallengeQueryAdapter to appropriate folder
